refactor: update page names to screen
