+++ src/screens/promoters/SupportCauseScreen/CommunityAddScreen/assets/community-add-cycle-pt.png
