+++ src/screens/promoters/SupportCauseScreen/CryptoScreen/SelectCryptoOfferSection/styles.ts
