+++ src/screens/promoters/SupportCauseScreen/CryptoScreen/styles.ts
