+++ src/screens/promoters/SupportCauseScreen/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { RootTabScreenProps } from "types";
-import CryptoPage from "screens/promoters/SupportCauseScreen/CryptoPage";
+import CryptoPage from "screens/promoters/SupportCauseScreen/CryptoScreen";
 import CardScreen from "screens/promoters/SupportCauseScreen/CardScreen";
 import { useCryptoPayment } from "contexts/cryptoPaymentContext";
